refactor(coding-agent): extract child lifecycle and usage owners
Refs ENG-5939, ENG-5938
