+++ packages/coding-agent/src/core/agent-session.ts
@@ -1,17 +1,17 @@
 import { AsyncLocalStorage } from "node:async_hooks";
 import { randomUUID } from "node:crypto";
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
-import { basename, dirname, join, resolve } from "node:path";
-import {
+import { dirname, resolve } from "node:path";
+import type {
 	Agent,
-	type AgentContext,
-	type AgentEvent,
-	type AgentMessage,
-	type AgentState,
-	type AgentTool,
-	type GetContinuationMessagesContext,
-	type ShouldStopAfterTurnContext,
-	type ThinkingLevel,
+	AgentContext,
+	AgentEvent,
+	AgentMessage,
+	AgentState,
+	AgentTool,
+	GetContinuationMessagesContext,
+	ShouldStopAfterTurnContext,
+	ThinkingLevel,
 } from "@earendil-works/pi-agent-core";
 import type {
 	Api,
@@ -40,6 +40,16 @@ import {
 	SessionBash,
 	type SessionBashEvent,
 } from "../session/bash.js";
+import { createChildSessionDir, createInlineChildRuntime } from "../session/child-runtime.js";
+import { SessionChildState } from "../session/child-state.js";
+import {
+	compactRlmText,
+	type RlmChildAgentSnapshot,
+	type RlmChildAgentStatus,
+	rlmChildLabel,
+} from "../session/child-types.js";
+import { SessionChildUsage } from "../session/child-usage.js";
+import { SessionChildren } from "../session/children.js";
 import { SessionCommitFence, type SessionCommitLease } from "../session/commit-fence.js";
 import { SessionCompaction, type SessionCompactionEvent } from "../session/compaction.js";
 import {
@@ -85,19 +95,13 @@ import { createTurnExecutionPolicy, type TurnExecutionPolicy, TurnPreparer } fro
 import { stripFrontmatter } from "../utils/frontmatter.js";
 import { waitForPromiseOrAbort } from "../utils/wait-for-abort.js";
 import {
-	AGENT_MESSAGE_CUSTOM_TYPE,
 	AGENT_MESSAGE_SKILL_NAME,
-	type AgentFamilyCatalogEntry,
 	type AgentSessionMessage,
-	type AgentSessionMessageAgentSummary,
 	type AgentSessionMessageController,
-	type AgentSessionMessageListResult,
 	type AgentSessionMessageReceipt,
 	assertAgentMessageQueueCapacity,
-	assertAgentSessionNameAvailable,
 	assertDirectAgentMessageTarget,
 	DEFAULT_AGENT_MESSAGE_MAX_PENDING_PER_SESSION,
-	formatAgentSessionNameUnavailable,
 	isAgentSessionMessage,
 	isAgentSessionMessagePrompt,
 	normalizeAgentSessionMessage,
@@ -201,8 +205,6 @@ import {
 	createAsyncBashCompletionMessage,
 	createHarnessDigestMessage,
 	createHeartbeatPromptMessage,
-	createRlmChildFailureMessage,
-	createRlmChildTerminalNoticeMessage,
 	createSessionSlashCommandMessage,
 	createSessionSlashCommandResultMessage,
 	HARNESS_DIGEST_CUSTOM_TYPE,
@@ -222,17 +224,12 @@ import { formatHarnessStateForPrompt, REFINE_SKILL_NAME, type RefinementResult }
 import type { ResourceLoader } from "./resource-loader.js";
 import {
 	type CreateRlmSubagentRuntimeOptions,
-	createDefaultRlmSubagentSessionName,
 	findRlmModelMatches,
-	normalizeRequestedRlmSubagentModel,
-	normalizeRequestedRlmSubagentSessionName,
-	normalizeRequestedRlmSubagentThinkingLevel,
 	type RlmCreateSessionResult,
 	type RlmDeleteSubagentResult,
 	type RlmFindModelsResult,
 	type RlmListSubagentsResult,
 	type RlmSpawnHandle,
-	type RlmSubagentRegistryEntry,
 	type RlmSubagentRuntime,
 	type SubagentRuntimeHost,
 } from "./rlm-runtime.js";
@@ -252,18 +249,12 @@ import {
 	type SessionActionSnapshot,
 	transitionSessionAction,
 } from "./session-action-store.js";
-import type {
-	BranchSummaryEntry,
-	ChildUsageAttributionEntry,
-	SessionContext,
-	SessionEntry,
-	SessionMessageEntry,
-} from "./session-manager.js";
+import type { BranchSummaryEntry, SessionContext, SessionEntry } from "./session-manager.js";
 import {
 	CURRENT_SESSION_VERSION,
 	getLatestCompactionEntry,
 	type SessionHeader,
-	SessionManager,
+	type SessionManager,
 } from "./session-manager.js";
 import type { SessionStats } from "./session-stats.js";
 import type { SettingsManager } from "./settings-manager.js";
@@ -277,46 +268,14 @@ import {
 import type { BuildSystemPromptOptions } from "./system-prompt.js";
 import { THINKING_LEVELS } from "./thinking-levels.js";
 import type { IpythonKernelProvisioner } from "./tools/ipython.js";
-import {
-	addAssistantUsage,
-	cloneUsage,
-	emptyUsage,
-	type SessionUsageSummary,
-	sessionUsageSummaryFrom,
-	subtractAssistantUsage,
-} from "./usage.js";
+import { emptyUsage, type SessionUsageSummary, sessionUsageSummaryFrom, subtractAssistantUsage } from "./usage.js";
 
+export type { RlmChildAgentActivity, RlmChildAgentSnapshot, RlmChildAgentStatus } from "../session/child-types.js";
+export { compactRlmText, rlmChildLabel } from "../session/child-types.js";
+export type { CompactionReason } from "../session/compaction.js";
 export type { GoalState, GoalStatus } from "./goals.js";
 export type { SessionStats } from "./session-stats.js";
 export { type ParsedSkillBlock, parseSkillBlock } from "./skill-blocks.js";
-
-export type RlmChildAgentStatus = "queued" | "running" | "done" | "error" | "cancelled";
-
-export interface RlmChildAgentActivity {
-	kind: "waiting" | "writing" | "executing";
-	toolName?: string;
-}
-
-export interface RlmChildAgentSnapshot {
-	id: string;
-	parentId?: string;
-	activeSessionId?: string;
-	sessionName?: string;
-	model?: string;
-	label: string;
-	status: RlmChildAgentStatus;
-	durationMs?: number;
-	answerPreview?: string;
-	toolUseCount?: number;
-	tokenCount?: number;
-	recap?: string;
-	sessionDir: string;
-	activity?: RlmChildAgentActivity;
-	repliedSinceTask?: boolean;
-	error?: string;
-}
-
-export type { CompactionReason } from "../session/compaction.js";
 
 export type AgentSessionEvent =
 	| AgentEvent
@@ -593,97 +552,17 @@ interface ModelSelectOptions {
 
 type AutonomousSlashCommand = { kind: "status" } | { kind: "on"; config?: AgentAutonomousConfig } | { kind: "off" };
 
-import type { RlmMaxDepthSource, RlmMaxDepthStatus, SetRlmMaxDepthResult } from "./rlm-max-depth.js";
+import type { RlmMaxDepthStatus, SetRlmMaxDepthResult } from "./rlm-max-depth.js";
 
 export type { RlmMaxDepthSource, RlmMaxDepthStatus, SetRlmMaxDepthResult } from "./rlm-max-depth.js";
-
-interface PersistedRlmMaxDepthState {
-	maxDepth: number;
-}
 
 type AutonomousRuntimeSnapshot = Pick<
 	AutonomousRuntimeState,
 	"continuationsUsed" | "gateAttempts" | "lastGateFailure" | "lastGateFailureSnapshot"
 >;
 
-interface RlmChildRun {
-	id: string;
-	prompt: string;
-	sessionName: string;
-	sessionDir: string;
-	model: Model<Api>;
-	status: RlmChildAgentStatus;
-	durationMs?: number;
-	answerPreview?: string;
-	toolUseCount: number;
-	activity?: RlmChildAgentActivity;
-	error?: string;
-	abort: () => void;
-	publication: AgentMessageDeferred;
-	/** Resolves after terminal result publication and detached-run cleanup finish. */
-	settlement: AgentMessageDeferred;
-	/** Child session, once its runtime exists. Used to cancel nested child runs. */
-	session?: AgentSession;
-	settled: boolean;
-	/** Do not inject a late terminal notice after the parent session is aborted. */
-	suppressTerminalNotice?: boolean;
-	/** Excluded from future strong barriers after an authoritative cancellation cut. */
-	abandonedForQuiescence?: boolean;
-	/** Selector snapshot for an admitted explicit delete. */
-	detachedDeletion?: RlmSubagentRegistryEntry;
-	/** Shared physical runtime cleanup owned by the explicit-delete path. */
-	deletionCleanup?: Promise<void>;
-	deletionCleanupObserver?: Promise<boolean>;
-	/** Resolves when a deletion may release its selector reservation. */
-	deletionReservation: AgentMessageDeferred;
-	deletionCleanupFailed?: boolean;
-	deletionRunFinished?: boolean;
-	deletionNotice?: Promise<void>;
-	deletionFailureNotice?: Promise<void>;
-	deletionNeedsCompletionNotice?: boolean;
-	completeDeletion?: () => Promise<void>;
-	reportDeletionCleanupFailure?: (error: unknown) => Promise<void>;
-	emitUpdate?: () => void;
-	lastEmittedUpdate?: string;
-	unsubscribe?: () => void;
-}
-
-interface RetainedRlmChild {
-	session: AgentSession;
-	run?: RlmChildRun;
-}
-
 interface RlmSubagentModelSelection {
 	model: Model<Api>;
-}
-
-const RLM_MAX_DEPTH_STATE_CUSTOM_TYPE = "rlm_max_depth_state";
-
-function noopRlmChildAbort(): void {}
-function noopRlmChildEventUnsubscribe(): void {}
-
-function isNonNegativeInteger(value: unknown): value is number {
-	return typeof value === "number" && Number.isSafeInteger(value) && value >= 0;
-}
-
-function parseDepth(value: string | undefined, fallback: number, name: string): number {
-	if (value === undefined || value === "") {
-		return fallback;
-	}
-	if (!/^\d+$/.test(value)) {
-		throw new Error(`${name} must be a non-negative integer`);
-	}
-	const parsed = Number(value);
-	if (!isNonNegativeInteger(parsed)) {
-		throw new Error(`${name} must be a non-negative integer`);
-	}
-	return parsed;
-}
-
-function isPersistedRlmMaxDepthState(value: unknown): value is PersistedRlmMaxDepthState {
-	return (
-		typeof value === "object" && value !== null && isNonNegativeInteger((value as PersistedRlmMaxDepthState).maxDepth)
-	);
 }
 
 const AUTONOMOUS_STATUS_NUMBER_FORMAT = new Intl.NumberFormat("en-US");
@@ -791,59 +670,6 @@ function parseAutonomousBudgetOptions(tokens: string[]): AgentAutonomousConfig {
 	return config;
 }
 
-export function compactRlmText(text: string, maxLength = 160): string {
-	const compact = text.replace(/\s+/g, " ").trim();
-	if (compact.length <= maxLength) {
-		return compact;
-	}
-	return `${compact.slice(0, Math.max(0, maxLength - 3)).trimEnd()}...`;
-}
-
-// Child-agent label: collapse to one line but keep the full prompt — the TUI
-// truncates to the visible width and elides shared prefixes, so capping here
-// would only hide the divergence between near-identical sibling prompts.
-export function rlmChildLabel(prompt: string): string {
-	return prompt.replace(/\s+/g, " ").trim() || "child agent";
-}
-
-function readAssistantText(message: AssistantMessage): string {
-	return message.content
-		.filter((block) => block.type === "text")
-		.map((block) => block.text)
-		.join("");
-}
-
-// Bounds how much accumulated child usage a parent process crash can lose.
-const RLM_CHILD_USAGE_FLUSH_MAX_PENDING_MS = 60_000;
-
-/** Label a child completion's usage by the nearest preceding prompt that triggered it. */
-function rlmChildUsageOrigin(
-	messages: readonly AgentMessage[],
-	assistant: AssistantMessage,
-): ChildUsageAttributionEntry["origin"] {
-	for (let index = messages.lastIndexOf(assistant) - 1; index >= 0; index--) {
-		const message = messages[index];
-		if (message.role !== "user" && message.role !== "custom") continue;
-		return message.role === "custom" && isAgentSessionMessage(message)
-			? message.details.id.startsWith("spawn:")
-				? "spawn_task"
-				: "agent_message"
-			: "direct_user";
-	}
-	return "direct_user";
-}
-
-function attributeChildUsage(parentUsage: Usage, childUsage: Usage): void {
-	const parentContextTokens =
-		parentUsage.totalTokens ||
-		parentUsage.input + parentUsage.output + parentUsage.cacheRead + parentUsage.cacheWrite;
-	// Recursive children are launched from an assistant tool call, so the parent assistant
-	// message carries their billable usage for session-level cost totals.
-	addAssistantUsage(parentUsage, childUsage);
-	// Child work affects session-level billable totals, not the parent's model-facing context size.
-	parentUsage.totalTokens = parentContextTokens;
-}
-
 export class AgentSession {
 	private readonly _tools: SessionTools;
 	private readonly _extensions: SessionExtensions;
@@ -876,6 +702,50 @@ export class AgentSession {
 	private get _baseSystemPromptOptions(): BuildSystemPromptOptions {
 		return this._tools.baseSystemPromptOptions;
 	}
+	private readonly _childState: SessionChildState;
+
+	private readonly _childUsage = new SessionChildUsage({
+		sessionManager: {
+			getEntries: () => this.sessionManager.getEntries(),
+			appendChildUsageAttribution: (...args) => this.sessionManager.appendChildUsageAttribution(...args),
+		},
+		invalidateOwnUsage: () => this._invalidateOwnUsage(),
+		afterParentDrain: (flush) => {
+			this._agentEventQueue = this._agentEventQueue.then(flush, flush);
+			this._agentEventQueue.catch(() => {});
+		},
+	});
+	private readonly _children = new SessionChildren({
+		isDisposed: () => this._disposed || this._disposing,
+		isInputSuspended: () => this._inputScheduler.suspended,
+		isStreaming: () => this.isStreaming,
+		isSessionActive: () => this.isSessionActive,
+		getMessageController: () => this._agentMessageController,
+		getDepth: () => this._childState.depth,
+		getMaxDepth: () => this._childState.maxDepth,
+		getParentNodeId: () => this._rlmParentNodeId,
+		getCwd: () => this._cwd,
+		getSessionId: () => this.sessionId,
+		getSessionName: () => this.sessionName,
+		getSessionFile: () => this.sessionFile,
+		getThinkingLevel: () => this.thinkingLevel,
+		getSemanticEdges: () => this._semanticEdges,
+		getChildOwner: (child) => child._children,
+		getParentReplyCount: (child) => child._childState.replyCount,
+		getChildSessionDir: (child) => child._rlmSessionDir,
+		resolveModel: (reference, target) => this._resolveRlmSubagentModel(reference, target),
+		createSessionDir: () => this._createChildRlmSessionDir(),
+		createRuntimeOptions: (request) => this._createRlmSubagentRuntimeOptions(request),
+		createRuntime: (options) => this._createRlmSubagentRuntime(options),
+		createUsageTracker: () => this._childUsage.createTracker(this._findLastAssistantMessage()),
+		hasDeferredTerminalNotices: () => this._hasDeferredRlmTerminalNotices(),
+		waitForHeadlessIdle: () => this.waitForHeadlessIdle(),
+		waitForActivityChange: (signal) => this._waitForSessionActivityChange(signal),
+		deliverTerminalNotice: (message) => this._deferRlmTerminalNotice(message),
+		emit: (event) => this._emit(event),
+		onSettled: () => this._maybeResumeGoalContinuationAfterRlmWork(),
+	});
+
 	private readonly _refinement: SessionRefinement;
 	readonly agent: Agent;
 	readonly sessionManager: SessionManager;
@@ -1083,44 +953,9 @@ export class AgentSession {
 	// re-populate the retained map after it's been cleared.
 	private _disposing = false;
 	private _disposeAsyncPromise?: Promise<void>;
-	private _rlmDepth: number;
-	private readonly _configuredRlmMaxDepth: number | undefined;
-	private _rlmMaxDepth: number;
-	private _rlmMaxDepthSource: RlmMaxDepthSource;
 	private readonly _semanticEdges: SemanticEdgeRecorder;
 	private _rlmParentNodeId?: string;
 	private _rlmParentAgent?: string;
-	private _repliedToParentSinceTask: boolean | undefined;
-	private _parentReplyCount = 0;
-	private _subagentRuntimeHost?: SubagentRuntimeHost;
-	// Shared by children charged to the same assistant; excludes usage not yet attributed on disk.
-	private _rlmDurableParentUsage = new WeakMap<AssistantMessage, Usage>();
-	// Child usage not yet represented by an indexed attribution, including a delayed parent entry.
-	private _rlmUnindexedChildUsage = new WeakMap<AssistantMessage, Usage>();
-	private _activeRlmChildRuns = new Map<string, RlmChildRun>();
-	private _unsettledRlmChildRuns = new Set<RlmChildRun>();
-	private _abandonedRlmQuiescenceChildIds = new Set<string>();
-	private _rlmQuiescenceWaitAborts = new Set<AbortController>();
-	private _pendingRlmSubagentSessionNames = new Set<string>();
-	// Inline mode keeps finished child sessions so the inspector can still read them;
-	// the daemon does the same by leaving the child session resident in its registry.
-	private _rlmChildSessions = new Map<string, RetainedRlmChild>();
-	private _deletedRlmChildIds = new Set<string>();
-	// Failed explicit deletes stay hidden from listings but retain their original
-	// selector so a later delete can retry cleanup without orphaning the runtime.
-	private _rlmChildCleanupFailures = new Map<string, RlmSubagentRegistryEntry>();
-	private _deletingRlmChildren = new Map<
-		string,
-		{
-			subagent: RlmSubagentRegistryEntry;
-			promise: Promise<RlmDeleteSubagentResult>;
-		}
-	>();
-	// Kept alive for retained children so nested updates (e.g. a grandchild cancel)
-	// still forward to root; torn down when the retained child is disposed.
-	private _rlmChildUnsubscribes = new Map<string, () => void>();
-	/** Latest recap for this session, written by the daemon summarizer; read by a parent to label its child snapshots. */
-	private _currentRecap?: string;
 
 	private _modelRegistry: ModelRegistry;
 
@@ -1169,7 +1004,7 @@ export class AgentSession {
 				isDisposing: () => this._disposing,
 				isStreaming: () => this.isStreaming,
 				isCompacting: () => this.isCompacting,
-				getDepth: () => this._rlmDepth,
+				getDepth: () => this._childState.depth,
 				getRlmSessionDir: () => this._rlmSessionDir,
 				getModel: () => this.model,
 				getThinkingLevel: () => this.thinkingLevel,
@@ -1208,17 +1043,21 @@ export class AgentSession {
 		this._agentMessageController = config.agentMessageController;
 		this._agentObserveController = config.agentObserveController;
 		this._mcpManager = config.mcpManager;
-		const headerRlmDepth = this.sessionManager.getHeader()?.rlmDepth;
-		this._rlmDepth =
-			config.rlmDepth ??
-			(isNonNegativeInteger(headerRlmDepth) ? headerRlmDepth : parseDepth(process.env.RLM_DEPTH, 0, "RLM_DEPTH"));
-		this._configuredRlmMaxDepth = config.rlmMaxDepth;
-		if (this._configuredRlmMaxDepth !== undefined && !isNonNegativeInteger(this._configuredRlmMaxDepth)) {
-			throw new Error("rlmMaxDepth must be a non-negative integer");
-		}
-		const resolvedRlmMaxDepth = this._resolveRlmMaxDepth();
-		this._rlmMaxDepth = resolvedRlmMaxDepth.maxDepth;
-		this._rlmMaxDepthSource = resolvedRlmMaxDepth.source;
+		this._childState = new SessionChildState(
+			{
+				sessionManager: this.sessionManager,
+				settingsManager: this.settingsManager,
+				refreshPrompt: (preserveExtensionPrompt) => {
+					const oldBase = this._baseSystemPrompt;
+					this._baseSystemPrompt = this._rebuildSystemPrompt(this.getActiveToolNames());
+					this.agent.state.systemPrompt = preserveExtensionPrompt
+						? this._refreshExtensionSystemPrompt(this.agent.state.systemPrompt, oldBase)
+						: this._baseSystemPrompt;
+				},
+				emitRecap: (recap) => this._emit({ type: "recap_update", recap }),
+			},
+			config,
+		);
 
 		this._rlmParentNodeId = config.rlmParentNodeId;
 		this._rlmParentAgent = config.rlmParentAgent;
@@ -1227,8 +1066,8 @@ export class AgentSession {
 				agentDir: this._agentDir,
 				authStorage: this._modelRegistry.authStorage,
 				resourceLoader: this._resourceLoader,
-				getDepth: () => this._rlmDepth,
-				getMaxDepth: () => this._rlmMaxDepth,
+				getDepth: () => this._childState.depth,
+				getMaxDepth: () => this._childState.maxDepth,
 				getArtifactDir: () => this.sessionManager.getSessionArtifactDir(),
 				getLocalHarnessStateDir: () => this._refinement._localHarnessStateDir(),
 			},
@@ -1249,7 +1088,7 @@ export class AgentSession {
 				emit: (event) => this._emit(event),
 				sendCustomMessage: (message, options) => this.sendCustomMessage(message, options),
 			},
-			(config.prewarmIpythonKernel ?? false) && this._rlmDepth === 0,
+			(config.prewarmIpythonKernel ?? false) && this._childState.depth === 0,
 		);
 		this._tools = new SessionTools(
 			{
@@ -1258,8 +1097,8 @@ export class AgentSession {
 				getExtensionRunner: () => this._extensionRunner,
 				getSessionFile: () => this.sessionManager.getSessionFile(),
 				getModelVisibleSkills: () => this._modelVisibleSkills(),
-				getDepth: () => this._rlmDepth,
-				getMaxDepth: () => this._rlmMaxDepth,
+				getDepth: () => this._childState.depth,
+				getMaxDepth: () => this._childState.maxDepth,
 				getParentAgent: () => this._rlmParentAgent,
 				getMcpManager: () => this._mcpManager,
 				getProvisioner: () => this._kernel.provisioner,
@@ -1332,13 +1171,8 @@ export class AgentSession {
 			spawnedByRequestId: config.semanticSpawnedByRequestId,
 		});
 		this.agent.streamFn = wrapStreamFnWithSemanticEdges(this.agent.streamFn, this._semanticEdges);
-		// A resumed child may have replied before this process started; false would
-		// claim knowledge that is not present in the session transcript.
-		this._repliedToParentSinceTask =
-			this._rlmDepth > 0 && this.sessionManager.getBranch().some((entry) => entry.type === "message")
-				? undefined
-				: false;
-		this._subagentRuntimeHost = config.subagentRuntimeHost;
+		this._childState.initializeParentReply();
+		this._children.setRuntimeHost(config.subagentRuntimeHost);
 		this._autonomousState = createAutonomousRuntimeState(config.autonomous, {
 			cwd: this._cwd,
 		});
@@ -1349,7 +1183,7 @@ export class AgentSession {
 		// thinking_level_change, service_tier_change) and no persisted
 		// thread_goal_state. This prevents reseeding after clear/complete/error
 		// or restart/rehydration of a session that already has messages or a goal.
-		if (this._rlmDepth === 0 && config.initialGoal && goalPersistence.canSeed()) {
+		if (this._childState.depth === 0 && config.initialGoal && goalPersistence.canSeed()) {
 			this._startGoal(config.initialGoal.objective, config.initialGoal.tokenBudget);
 			// Goal context is the model's only source of goal visibility; action
 			// admission is unavailable mid-construction, so ride the next turn.
@@ -1406,7 +1240,7 @@ export class AgentSession {
 	}
 
 	setSubagentRuntimeHost(host?: SubagentRuntimeHost): void {
-		this._subagentRuntimeHost = host;
+		this._children.setRuntimeHost(host);
 	}
 
 	private async _getRequiredRequestAuth(model: Model<any>): Promise<{
@@ -1527,52 +1361,8 @@ export class AgentSession {
 		this._emit({ type: "goal_update", goal: this.goalState });
 	}
 
-	private _loadPersistedRlmMaxDepthState(): PersistedRlmMaxDepthState | undefined {
-		const branch = this.sessionManager.getBranch();
-		for (let i = branch.length - 1; i >= 0; i--) {
-			const entry = branch[i];
-			if (
-				entry.type === "custom" &&
-				entry.customType === RLM_MAX_DEPTH_STATE_CUSTOM_TYPE &&
-				isPersistedRlmMaxDepthState(entry.data)
-			) {
-				return entry.data;
-			}
-		}
-		return undefined;
-	}
-
-	private _resolveRlmMaxDepth(): {
-		maxDepth: number;
-		source: RlmMaxDepthSource;
-	} {
-		const persisted = this._loadPersistedRlmMaxDepthState();
-		if (persisted) {
-			return { maxDepth: persisted.maxDepth, source: "chat" };
-		}
-		if (this._configuredRlmMaxDepth !== undefined) {
-			return { maxDepth: this._configuredRlmMaxDepth, source: "inherited" };
-		}
-		const global = this.settingsManager.getRlmMaxDepth();
-		if (global !== undefined && isNonNegativeInteger(global)) {
-			return { maxDepth: global, source: "global" };
-		}
-		const env = process.env.RLM_MAX_DEPTH;
-		if (env !== undefined && env !== "") {
-			return { maxDepth: parseDepth(env, 1, "RLM_MAX_DEPTH"), source: "env" };
-		}
-		return { maxDepth: 2, source: "default" };
-	}
-
 	private _reloadRlmMaxDepthFromBranch(): void {
-		const previousMaxDepth = this._rlmMaxDepth;
-		const resolved = this._resolveRlmMaxDepth();
-		this._rlmMaxDepth = resolved.maxDepth;
-		this._rlmMaxDepthSource = resolved.source;
-		if (resolved.maxDepth !== previousMaxDepth) {
-			this._baseSystemPrompt = this._rebuildSystemPrompt(this.getActiveToolNames());
-			this.agent.state.systemPrompt = this._baseSystemPrompt;
-		}
+		this._childState.reloadFromBranch();
 	}
 
 	private _cancelSessionActions(
@@ -3009,34 +2799,7 @@ export class AgentSession {
 	private async _disposeAsyncOnce(kernelSnapshot: boolean): Promise<void> {
 		// Flush kernels/traces for both still-running and retained children; the sync
 		// dispose() below only tears them down synchronously.
-		for (const run of [...this._activeRlmChildRuns.values()]) {
-			const childSession = run.session;
-			if (!childSession) continue;
-			if (run.detachedDeletion) {
-				run.suppressTerminalNotice = true;
-				if (run.deletionCleanupObserver) {
-					await run.deletionCleanupObserver.catch(() => false);
-				} else if (run.deletionCleanup) {
-					await run.deletionCleanup.catch(() => childSession.disposeAsync().catch(() => undefined));
-				} else {
-					// Cleanup already failed and was exposed for retry before disposal.
-					await childSession.disposeAsync().catch(() => undefined);
-				}
-				if (!run.settled) await this._finishRlmRunDeletion(run);
-			} else {
-				await childSession.disposeAsync().catch(() => undefined);
-			}
-		}
-		for (const unsubscribe of this._rlmChildUnsubscribes.values()) {
-			unsubscribe();
-		}
-		this._rlmChildUnsubscribes.clear();
-		for (const { session } of this._rlmChildSessions.values()) {
-			await session.disposeAsync().catch(() => undefined);
-		}
-		this._rlmChildSessions.clear();
-		this._rlmChildCleanupFailures.clear();
-		this._deletedRlmChildIds.clear();
+		await this._children.disposeAsync();
 		await this._kernel.dispose(kernelSnapshot);
 		this.dispose();
 		await this._disposeCallbacksPromise;
@@ -3067,24 +2830,13 @@ export class AgentSession {
 			return;
 		}
 		this._disposed = true;
-		for (const run of this._unsettledRlmChildRuns) run.suppressTerminalNotice = true;
-		for (const controller of this._rlmQuiescenceWaitAborts) controller.abort();
+		this._children.beginDisposal();
 		this._commitFence.dispose();
 		try {
 			// Invalidate scheduled timers and abort any in-flight review so a late
 			// resolution cannot write harness state or re-subscribe handlers.
 			this._refinement.dispose();
-			this._cancelActiveRlmChildRuns("Parent session disposed");
-			for (const unsubscribe of this._rlmChildUnsubscribes.values()) {
-				unsubscribe();
-			}
-			this._rlmChildUnsubscribes.clear();
-			for (const { session } of this._rlmChildSessions.values()) {
-				session.dispose();
-			}
-			this._rlmChildSessions.clear();
-			this._rlmChildCleanupFailures.clear();
-			this._deletedRlmChildIds.clear();
+			this._children.dispose();
 			this._pendingNextTurnMessages = [];
 			const deliveryError = new Error("Session disposed before prompt delivery.");
 			const completionError = new Error("Session disposed before prompt completion.");
@@ -3207,7 +2959,7 @@ export class AgentSession {
 	}
 
 	get rlmDepth(): number {
-		return this._rlmDepth;
+		return this._childState.depth;
 	}
 
 	get semanticEdges(): SemanticEdgeRecorder {
@@ -3215,7 +2967,7 @@ export class AgentSession {
 	}
 
 	get rlmMaxDepth(): number {
-		return this._rlmMaxDepth;
+		return this._childState.maxDepth;
 	}
 
 	get sessionName(): string | undefined {
@@ -3439,7 +3191,7 @@ export class AgentSession {
 			customMessage,
 			admissionCommitted,
 		});
-		if (customMessage?.details.fromRelationship === "parent") this._repliedToParentSinceTask = false;
+		if (customMessage?.details.fromRelationship === "parent") this._childState.resetReply();
 	}
 
 	async queueAgentMessagePrompt(
@@ -3453,14 +3205,14 @@ export class AgentSession {
 				agentMessageId,
 				message: customMessage,
 			});
-			if (customMessage?.details.fromRelationship === "parent") this._repliedToParentSinceTask = false;
+			if (customMessage?.details.fromRelationship === "parent") this._childState.resetReply();
 			return true;
 		}
 		const queued = await this._queuePreparedPrompt("followUp", text, undefined, {
 			agentMessageId,
 			message: customMessage,
 		});
-		if (queued && customMessage?.details.fromRelationship === "parent") this._repliedToParentSinceTask = false;
+		if (queued && customMessage?.details.fromRelationship === "parent") this._childState.resetReply();
 		return queued;
 	}
 
@@ -5435,10 +5187,7 @@ export class AgentSession {
 	}
 
 	requestAbort(): void {
-		for (const run of [...this._unsettledRlmChildRuns]) {
-			if (run.status === "cancelled") this._abandonRlmRunForQuiescence(run);
-		}
-		for (const controller of this._rlmQuiescenceWaitAborts) controller.abort();
+		this._children.requestAbort();
 		this._inputScheduler.suspend("abort");
 		this._demoteRlmTerminalNoticeActions();
 		this._cancelSessionActions(
@@ -5481,7 +5230,7 @@ export class AgentSession {
 		this._inputScheduler.suspend("update-restart");
 		this._cancelPostCompactionContinue();
 		this.abortRetry();
-		for (const controller of this._rlmQuiescenceWaitAborts) controller.abort();
+		this._children.cancelQuiescenceWaits();
 		this._cancelActiveRlmChildRuns("Parent session aborted for update restart");
 		this._goalAbortInProgress = this._goals.state.status === "active";
 		this.agent.abort();
@@ -5831,11 +5580,8 @@ export class AgentSession {
 		return performSessionCompaction(this._compactionExecution, options);
 	}
 
-	private async _reapDeletedRlmSubagentRuntimesAfterCompaction(): Promise<void> {
-		const childIds = [...this._rlmChildCleanupFailures.keys()].filter(
-			(childId) => !this._activeRlmChildRuns.get(childId)?.detachedDeletion,
-		);
-		await Promise.allSettled(childIds.map((childId) => this.deleteRlmSubagent(childId)));
+	private _reapDeletedRlmSubagentRuntimesAfterCompaction(): Promise<void> {
+		return this._children.reapAfterCompaction();
 	}
 
 	abortCompaction(): void {
@@ -6062,12 +5808,9 @@ export class AgentSession {
 			getAgentMessageController: () => this._agentMessageController,
 			hasObserveController: () => !!this._agentObserveController,
 			getMcpManager: () => this._mcpManager,
-			getDepth: () => this._rlmDepth,
+			getDepth: () => this._childState.depth,
 			awaitChildPublication: (selector) => this._awaitPendingRlmChildPublication(selector),
-			recordParentReply: () => {
-				this._repliedToParentSinceTask = true;
-				this._parentReplyCount += 1;
-			},
+			recordParentReply: () => this._childState.recordReply(),
 			handleGoal: (type, payload) => this.handleGoalHostRequest(type, payload),
 			handleCompact: (type, payload) => this.handleCompactHostRequest(type, payload),
 			handleRefine: (type, payload) => this.handleRefineHostRequest(type, payload),
@@ -6112,20 +5855,7 @@ export class AgentSession {
 	// does RLM work. The temp dir is created lazily in _createChildRlmSessionDir.
 
 	private _createChildRlmSessionDir(): string {
-		const parentDir = this._ensureRlmSessionDir() ?? this._createEphemeralRlmSessionDir();
-		for (let i = 0; i < 100; i++) {
-			const childDir = join(parentDir, `sub-${randomUUID().slice(0, 8)}`);
-			try {
-				mkdirSync(childDir);
-				return childDir;
-			} catch (error) {
-				if (error instanceof Error && "code" in error && error.code === "EEXIST") {
-					continue;
-				}
-				throw error;
-			}
-		}
-		throw new Error("Unable to create unique RLM child session directory");
+		return createChildSessionDir(() => this._ensureRlmSessionDir() ?? this._createEphemeralRlmSessionDir());
 	}
 
 	private _rlmKernelEnv(): Record<string, string> {
@@ -6144,23 +5874,15 @@ export class AgentSession {
 	}
 
 	setCurrentRecap(recap: string | undefined): void {
-		if (this._currentRecap === recap) return;
-		this._currentRecap = recap;
-		this._emit({ type: "recap_update", recap });
+		this._childState.setCurrentRecap(recap);
 	}
 
 	get repliedToParentSinceTask(): boolean | undefined {
-		return this._repliedToParentSinceTask;
+		return this._childState.repliedSinceTask;
 	}
 
 	getCurrentRecap(): string | undefined {
-		return this._currentRecap;
-	}
-
-	private _findAssistantEntryForMessage(message: AssistantMessage): SessionMessageEntry | undefined {
-		return this.sessionManager
-			.getEntries()
-			.find((entry): entry is SessionMessageEntry => entry.type === "message" && entry.message === message);
+		return this._childState.getCurrentRecap();
 	}
 
 	private _createRlmSubagentRuntimeOptions(options: {
@@ -6191,553 +5913,61 @@ export class AgentSession {
 			customTools: [...this._customTools],
 			includeGoals: this._includeGoals,
 			includeCompactSkill: this._includeCompactSkill,
-			rlmDepth: this._rlmDepth + 1,
-			rlmMaxDepth: this._rlmMaxDepth,
+			rlmDepth: this._childState.depth + 1,
+			rlmMaxDepth: this._childState.maxDepth,
 			rlmParentNodeId: options.id,
 			spawnedByRequestId: options.spawnedByRequestId,
 		};
 	}
 
 	private async _createRlmSubagentRuntime(options: CreateRlmSubagentRuntimeOptions): Promise<RlmSubagentRuntime> {
-		if (this._subagentRuntimeHost) {
-			return await this._subagentRuntimeHost.createRlmSubagentRuntime(options);
+		const host = this._children.getRuntimeHost();
+		if (host) {
+			return await host.createRlmSubagentRuntime(options);
 		}
 
 		return this._createInlineRlmSubagentRuntime(options);
 	}
 
 	private _createInlineRlmSubagentRuntime(options: CreateRlmSubagentRuntimeOptions): RlmSubagentRuntime {
-		const childSessionManager = SessionManager.create(this._cwd, options.sessionDir);
-		if (options.parentSession.sessionFile) {
-			childSessionManager.newSession({
-				parentSession: options.parentSession.sessionFile,
-				rlmDepth: options.rlmDepth,
-			});
-		}
-		childSessionManager.appendModelChange(options.model.provider, options.model.id);
-		childSessionManager.appendThinkingLevelChange(options.thinkingLevel);
-		childSessionManager.appendServiceTierChange(options.serviceTier);
-
-		const childAgent = new Agent({
-			initialState: {
-				systemPrompt: "",
-				model: options.model,
-				thinkingLevel: options.thinkingLevel,
-				serviceTier: options.serviceTier,
-				tools: [],
+		return createInlineChildRuntime(
+			{
+				cwd: this._cwd,
+				agentDir: this._agentDir,
+				agent: this.agent,
+				settingsManager: this.settingsManager,
+				resourceLoader: this._resourceLoader,
+				modelRegistry: this._modelRegistry,
 			},
-			convertToLlm: this.agent.convertToLlm,
-			transformContext: this.agent.transformContext,
-			streamFn: this.agent.streamFn,
-			getApiKey: this.agent.getApiKey,
-			onPayload: this.agent.onPayload,
-			onResponse: this.agent.onResponse,
-			steeringMode: this.settingsManager.getSteeringMode(),
-			followUpMode: this.settingsManager.getFollowUpMode(),
-			sessionId: childSessionManager.getSessionId(),
-			thinkingBudgets: this.settingsManager.getThinkingBudgets(),
-			transport: this.settingsManager.getTransport(),
-			toolExecution: this.agent.toolExecution,
-		});
-
-		const child = new AgentSession({
-			agent: childAgent,
-			sessionManager: childSessionManager,
-			settingsManager: this.settingsManager,
-			cwd: this._cwd,
-			agentDir: this._agentDir,
-			scopedModels: options.scopedModels,
-			resourceLoader: this._resourceLoader,
-			customTools: options.customTools,
-			modelRegistry: this._modelRegistry,
-			initialActiveToolNames: options.activeToolNames,
-			allowedToolNames: options.allowedToolNames,
-			includeGoals: options.includeGoals,
-			includeCompactSkill: options.includeCompactSkill,
-			rlmDepth: options.rlmDepth,
-			rlmMaxDepth: options.rlmMaxDepth,
-			rlmSessionDir: options.sessionDir,
-			rlmParentNodeId: options.rlmParentNodeId,
-			rlmParentAgent: options.parentSession.sessionName ?? options.parentSession.sessionId,
-			semanticParentSessionId: options.parentSession.sessionId,
-			semanticSpawnedByRequestId: options.spawnedByRequestId,
-			sessionStartEvent: { type: "session_start", reason: "startup" },
-		});
-		if (child.sessionName !== options.sessionName) {
-			try {
-				child.setSessionName(options.sessionName);
-			} catch (error) {
-				child.dispose();
-				throw error;
-			}
-		}
-		options.onSessionPublished?.(child);
-
-		return { session: child };
-	}
-
-	private _abandonRlmRunForQuiescence(run: RlmChildRun): void {
-		run.suppressTerminalNotice = true;
-		run.abandonedForQuiescence = true;
-		this._abandonedRlmQuiescenceChildIds.add(run.id);
-		this._unsettledRlmChildRuns.delete(run);
-		run.settlement.resolve();
-		this._maybeResumeGoalContinuationAfterRlmWork();
+			options,
+		);
 	}
 
 	private _cancelActiveRlmChildRuns(reason: string): void {
-		for (const run of this._activeRlmChildRuns.values()) {
-			this._cancelRlmChildRun(run, reason);
-		}
-	}
-
-	private _cancelRlmChildRun(run: RlmChildRun, reason: string): boolean {
-		if (run.status !== "running" && run.status !== "queued") {
-			return false;
-		}
-		run.status = "cancelled";
-		if (this._inputScheduler.suspended) this._abandonRlmRunForQuiescence(run);
-		run.error = reason;
-		run.publication.reject(new Error(reason));
-		run.abort();
-		// Surface the cancellation immediately; the run's own terminal update is
-		// delayed indefinitely when the child is stuck mid-stream, which is
-		// exactly when users reach for the kill.
-		run.emitUpdate?.();
-		return true;
+		this._children.cancelActiveRuns(reason);
 	}
 
 	getRlmChildRunStatus(childId: string): RlmChildAgentStatus | undefined {
-		return this._activeRlmChildRuns.get(childId)?.status;
+		return this._children.getRlmChildRunStatus(childId);
 	}
 
-	private async _currentActiveSessionId(): Promise<string | undefined> {
-		try {
-			return (await this._agentMessageController?.listAgents())?.current?.activeSessionId;
-		} catch {
-			return undefined;
-		}
+	private _awaitPendingRlmChildPublication(selector: string): Promise<string | undefined> {
+		return this._children.awaitPublication(selector);
 	}
 
-	private async _awaitPendingRlmChildPublication(selector: string): Promise<string | undefined> {
-		const run = [...this._activeRlmChildRuns.values()].find(
-			(candidate) =>
-				(candidate.status === "queued" || candidate.status === "running" || candidate.status === "done") &&
-				!candidate.detachedDeletion &&
-				(candidate.id === selector || candidate.sessionName === selector),
-		);
-		if (!run) return undefined;
-		await run.publication.promise;
-		return run.session?.sessionId;
+	listRlmSubagents(): Promise<RlmListSubagentsResult> {
+		return this._children.listRlmSubagents();
 	}
 
-	async listRlmSubagents(): Promise<RlmListSubagentsResult> {
-		return this._buildRlmSubagentList(await this._agentMessageController?.listAgents());
-	}
-
-	private _buildRlmSubagentList(listedAgents?: AgentSessionMessageListResult): RlmListSubagentsResult {
-		const daemonChildren = new Map<string, AgentSessionMessageAgentSummary>();
-		const parentActiveSessionId = listedAgents?.current?.activeSessionId;
-		if (parentActiveSessionId) {
-			for (const agent of listedAgents.agents) {
-				if (
-					agent.runtimeKind === "subagent" &&
-					agent.parentActiveSessionId === parentActiveSessionId &&
-					agent.rlmChildId
-				) {
-					daemonChildren.set(agent.rlmChildId, agent);
-				}
-			}
-		}
-
-		const subagents: RlmListSubagentsResult["subagents"] = [];
-		const recorded = new Set<string>();
-		for (const run of this._activeRlmChildRuns.values()) {
-			if (this._deletingRlmChildren.has(run.id) || run.detachedDeletion || run.status === "cancelled") {
-				continue;
-			}
-			const daemonChild = daemonChildren.get(run.id);
-			subagents.push({
-				rlm_child_id: run.id,
-				active_session_id: daemonChild?.activeSessionId ?? null,
-				session_id: daemonChild?.sessionId ?? run.session?.sessionId ?? null,
-				session_name: daemonChild?.sessionName ?? run.session?.sessionName ?? run.sessionName,
-				session_dir: run.sessionDir,
-				status: run.status === "done" ? "completed" : run.status === "error" ? "error" : "running",
-			});
-			recorded.add(run.id);
-		}
-		for (const [childId, { session: childSession }] of this._rlmChildSessions) {
-			if (
-				this._deletingRlmChildren.has(childId) ||
-				recorded.has(childId) ||
-				this._rlmChildCleanupFailures.has(childId)
-			) {
-				continue;
-			}
-			const daemonChild = daemonChildren.get(childId);
-			const sessionDir = childSession._rlmSessionDir;
-			if (!sessionDir) {
-				continue;
-			}
-			subagents.push({
-				rlm_child_id: childId,
-				active_session_id: daemonChild?.activeSessionId ?? null,
-				session_id: daemonChild?.sessionId ?? childSession.sessionId,
-				session_name:
-					daemonChild?.sessionName ?? childSession.sessionName ?? createDefaultRlmSubagentSessionName("", childId),
-				session_dir: sessionDir,
-				status: "completed",
-			});
-			recorded.add(childId);
-		}
-		for (const [childId, daemonChild] of daemonChildren) {
-			if (
-				recorded.has(childId) ||
-				this._deletingRlmChildren.has(childId) ||
-				this._deletedRlmChildIds.has(childId) ||
-				this._rlmChildCleanupFailures.has(childId) ||
-				!daemonChild.sessionDir
-			) {
-				continue;
-			}
-			subagents.push({
-				rlm_child_id: childId,
-				active_session_id: daemonChild.activeSessionId,
-				session_id: daemonChild.sessionId,
-				session_name: daemonChild.sessionName ?? createDefaultRlmSubagentSessionName("", childId),
-				session_dir: daemonChild.sessionDir,
-				status: daemonChild.rlmChildRegistryStatus === "completed" ? "completed" : "error",
-			});
-		}
-		return { subagents };
-	}
-
-	private _rlmSubagentMatchesTarget(entry: RlmSubagentRegistryEntry, target: string): boolean {
-		return (
-			entry.rlm_child_id === target ||
-			entry.active_session_id === target ||
-			entry.session_id === target ||
-			entry.session_name === target
-		);
-	}
-
-	private async _resolveDirectRlmSubagent(target: string): Promise<RlmSubagentRegistryEntry> {
-		const candidates = [...(await this.listRlmSubagents()).subagents, ...this._rlmChildCleanupFailures.values()];
-		const matches = candidates.filter((entry) => this._rlmSubagentMatchesTarget(entry, target));
-		if (matches.length === 0) {
-			throw new Error(`No direct RLM subagent matches "${target}" in the current parent session`);
-		}
-		if (matches.length > 1) {
-			throw new Error(`RLM subagent selector "${target}" is ambiguous in the current parent session`);
-		}
-		return matches[0]!;
-	}
-
-	async deleteInactiveRlmSubagent(
+	deleteInactiveRlmSubagent(
 		childId: string,
 		isExternallyRunning: () => boolean = () => false,
 	): Promise<"deleted" | "not_found" | "running"> {
-		for (const owner of this._rlmSubtreeSessions()) {
-			const isRunning = (): boolean => {
-				const status = owner._activeRlmChildRuns.get(childId)?.status;
-				return status === "queued" || status === "running" || isExternallyRunning();
-			};
-			if (isRunning()) {
-				return "running";
-			}
-			const subagent = [
-				...(await owner.listRlmSubagents()).subagents,
-				...owner._rlmChildCleanupFailures.values(),
-			].find((entry) => entry.rlm_child_id === childId);
-			if (!subagent) continue;
-			if (isRunning()) {
-				return "running";
-			}
-			const result = await owner._trackRlmSubagentDeletion(subagent, () => {
-				if (isRunning()) {
-					return Promise.resolve({ subagent, outcome: "skipped_running" });
-				}
-				return owner._deleteResolvedRlmSubagent(subagent);
-			});
-			return result.outcome === "skipped_running" ? "running" : "deleted";
-		}
-		return "not_found";
+		return this._children.deleteInactiveRlmSubagent(childId, isExternallyRunning);
 	}
 
-	async deleteRlmSubagent(target: string): Promise<RlmDeleteSubagentResult> {
-		const inFlight = [...this._deletingRlmChildren.values()].filter(({ subagent }) =>
-			this._rlmSubagentMatchesTarget(subagent, target),
-		);
-		if (inFlight.length > 1) {
-			throw new Error(`RLM subagent selector "${target}" is ambiguous in the current parent session`);
-		}
-
-		// Running and retained children can be reserved synchronously. This keeps
-		// them hidden immediately while the async daemon listing checks for a
-		// conflicting passive selector.
-		const localMatches = [
-			...this._buildRlmSubagentList().subagents,
-			...this._rlmChildCleanupFailures.values(),
-		].filter((entry) => this._rlmSubagentMatchesTarget(entry, target));
-		const matchingChildIds = new Set([
-			...inFlight.map(({ subagent }) => subagent.rlm_child_id),
-			...localMatches.map((subagent) => subagent.rlm_child_id),
-		]);
-		if (matchingChildIds.size > 1 || localMatches.length > 1) {
-			throw new Error(`RLM subagent selector "${target}" is ambiguous in the current parent session`);
-		}
-		if (inFlight[0]) {
-			return inFlight[0].promise;
-		}
-		if (localMatches[0]) {
-			const subagent = localMatches[0];
-			return this._trackRlmSubagentDeletion(subagent, async () => {
-				const listedAgents = await this._agentMessageController?.listAgents();
-				const listedSubagents = this._buildRlmSubagentList(listedAgents).subagents;
-				const passiveMatches = listedSubagents.filter(
-					(entry) => entry.rlm_child_id !== subagent.rlm_child_id && this._rlmSubagentMatchesTarget(entry, target),
-				);
-				if (passiveMatches.length > 0) {
-					throw new Error(`RLM subagent selector "${target}" is ambiguous in the current parent session`);
-				}
-				const parentActiveSessionId = listedAgents?.current?.activeSessionId;
-				const daemonChild = listedAgents?.agents.find(
-					(agent) =>
-						agent.rlmChildId === subagent.rlm_child_id && agent.parentActiveSessionId === parentActiveSessionId,
-				);
-				const resolvedSubagent = daemonChild
-					? {
-							...subagent,
-							active_session_id: daemonChild.activeSessionId,
-							session_id: daemonChild.sessionId,
-							session_name: daemonChild.sessionName ?? subagent.session_name,
-						}
-					: subagent;
-				return this._deleteResolvedRlmSubagent(resolvedSubagent);
-			});
-		}
-
-		const directMatches = [
-			...(await this.listRlmSubagents()).subagents,
-			...this._rlmChildCleanupFailures.values(),
-		].filter((entry) => this._rlmSubagentMatchesTarget(entry, target));
-		const directChildIds = new Set(directMatches.map((subagent) => subagent.rlm_child_id));
-		if (directChildIds.size > 1) {
-			throw new Error(`RLM subagent selector "${target}" is ambiguous in the current parent session`);
-		}
-		const subagent = directMatches[0] ?? (await this._resolveDirectRlmSubagent(target));
-		return this._trackRlmSubagentDeletion(subagent, () => this._deleteResolvedRlmSubagent(subagent));
-	}
-
-	private async _trackRlmSubagentDeletion(
-		subagent: RlmSubagentRegistryEntry,
-		startDeletion: () => Promise<RlmDeleteSubagentResult>,
-	): Promise<RlmDeleteSubagentResult> {
-		const existing = this._deletingRlmChildren.get(subagent.rlm_child_id);
-		if (existing) return existing.promise;
-		const deletion = Promise.resolve().then(startDeletion);
-		this._deletingRlmChildren.set(subagent.rlm_child_id, {
-			subagent,
-			promise: deletion,
-		});
-		try {
-			return await deletion;
-		} finally {
-			const clearReservation = () => {
-				if (this._deletingRlmChildren.get(subagent.rlm_child_id)?.promise === deletion) {
-					this._deletingRlmChildren.delete(subagent.rlm_child_id);
-				}
-			};
-			const run = this._activeRlmChildRuns.get(subagent.rlm_child_id);
-			if (run?.detachedDeletion) {
-				// Keep every selector reserved until the run settles, or until a failed
-				// cleanup is exposed for an explicit retry. Repeated deletes before that
-				// boundary return the same accepted result.
-				void run.deletionReservation.promise.then(clearReservation, clearReservation);
-			} else {
-				clearReservation();
-			}
-		}
-	}
-
-	private _deleteRlmSubagentSession(childId: string, session?: AgentSession): Promise<void> {
-		if (this._subagentRuntimeHost) {
-			return this._subagentRuntimeHost.deleteRlmSubagentRuntime(childId, session);
-		}
-		return session?.disposeAsync() ?? Promise.resolve();
-	}
-
-	private _ensureRlmRunDeletionCleanup(run: RlmChildRun, session: AgentSession): Promise<void> {
-		if (run.deletionCleanup) return run.deletionCleanup;
-		const cleanup = Promise.resolve().then(() => this._deleteRlmSubagentSession(run.id, session));
-		run.deletionCleanup = cleanup;
-		// Deletion admission is intentionally nonblocking. The detached run owner
-		// joins this exact promise before settlement and records any failure.
-		void cleanup.catch(() => undefined);
-		return cleanup;
-	}
-
-	private async _recordRlmRunDeletionCleanupFailure(
-		run: RlmChildRun,
-		subagent: RlmSubagentRegistryEntry,
-		session: AgentSession,
-		error: unknown,
-	): Promise<void> {
-		if (this._disposed || this._disposing) {
-			run.suppressTerminalNotice = true;
-			await session.disposeAsync().catch(() => undefined);
-			if (!run.settled) await this._finishRlmRunDeletion(run);
-			return;
-		}
-		run.deletionCleanup = undefined;
-		run.deletionCleanupObserver = undefined;
-		run.deletionCleanupFailed = true;
-		run.session = session;
-		this._rlmChildCleanupFailures.set(run.id, subagent);
-		// Make retry admission available before waking the parent model with the
-		// retry-required notice.
-		run.deletionReservation.resolve();
-		await Promise.resolve();
-		await run.reportDeletionCleanupFailure?.(error);
-	}
-
-	private async _finishRlmRunDeletion(run: RlmChildRun): Promise<void> {
-		await run.completeDeletion?.();
-		if (this._activeRlmChildRuns.get(run.id) === run) {
-			this._removeRlmSubagentTracking(run.id, run);
-		}
-		run.settled = true;
-		run.settlement.resolve();
-		run.deletionReservation.resolve();
-		this._unsettledRlmChildRuns.delete(run);
-		this._maybeResumeGoalContinuationAfterRlmWork();
-	}
-
-	private _observeRlmRunDeletionCleanup(
-		run: RlmChildRun,
-		subagent: RlmSubagentRegistryEntry,
-		session: AgentSession,
-		cleanup: Promise<void>,
-	): Promise<boolean> {
-		if (run.deletionCleanupObserver) return run.deletionCleanupObserver;
-		const observer = cleanup.then(
-			() => true,
-			async (error) => {
-				await this._recordRlmRunDeletionCleanupFailure(run, subagent, session, error);
-				return false;
-			},
-		);
-		run.deletionCleanupObserver = observer;
-		void observer.catch(() => undefined);
-		return observer;
-	}
-
-	private _continueFinishedRlmRunDeletion(
-		run: RlmChildRun,
-		subagent: RlmSubagentRegistryEntry,
-		session: AgentSession,
-	): void {
-		const cleanup = this._ensureRlmRunDeletionCleanup(run, session);
-		const observer = this._observeRlmRunDeletionCleanup(run, subagent, session, cleanup);
-		if (!run.deletionRunFinished) return;
-		void observer
-			.then(async (cleanupSucceeded) => {
-				if (cleanupSucceeded) await this._finishRlmRunDeletion(run);
-			})
-			.catch(() => undefined);
-	}
-
-	private _removeRlmSubagentTracking(childId: string, run?: RlmChildRun): void {
-		run?.unsubscribe?.();
-		this._rlmChildUnsubscribes.get(childId)?.();
-		this._rlmChildUnsubscribes.delete(childId);
-		this._rlmChildSessions.delete(childId);
-		this._rlmChildCleanupFailures.delete(childId);
-		this._abandonedRlmQuiescenceChildIds.delete(childId);
-		if (!run || this._activeRlmChildRuns.get(childId) === run) {
-			this._activeRlmChildRuns.delete(childId);
-		}
-		if (run) {
-			run.abort = noopRlmChildAbort;
-			run.unsubscribe = undefined;
-			run.session = undefined;
-		}
-	}
-
-	private _emitRlmSubagentRemoval(subagent: RlmSubagentRegistryEntry): void {
-		this._emit({
-			type: "rlm_child_update",
-			child: {
-				id: subagent.rlm_child_id,
-				parentId: this._rlmParentNodeId,
-				activeSessionId: subagent.active_session_id ?? undefined,
-				sessionName: subagent.session_name,
-				label: subagent.session_name,
-				status: "cancelled",
-				sessionDir: subagent.session_dir,
-				error: "Deleted by parent orchestrator",
-			},
-		});
-	}
-
-	private async _deleteResolvedRlmSubagent(subagent: RlmSubagentRegistryEntry): Promise<RlmDeleteSubagentResult> {
-		const childId = subagent.rlm_child_id;
-		const run = this._activeRlmChildRuns.get(childId);
-		if (run) {
-			if (run.deletionCleanupFailed) {
-				// Reset retry coordination only after selector preflight reaches the
-				// resolved child. A failed preflight must leave the prior retry boundary
-				// intact so a later call can acquire it.
-				run.deletionCleanupFailed = false;
-				run.deletionFailureNotice = undefined;
-				run.deletionReservation = createAgentMessageDeferred();
-			}
-			// The detached task remains the sole lifecycle owner. Mark deletion before
-			// cancellation so its catch/finally path cannot race a normal release or
-			// terminal notice against the physical delete.
-			run.detachedDeletion = subagent;
-			if (this._cancelRlmChildRun(run, "Deleted by parent orchestrator")) {
-				run.deletionNeedsCompletionNotice = true;
-			} else {
-				this._emitRlmSubagentRemoval(subagent);
-			}
-			const liveSession = run.session;
-			if (run.status === "error" && !liveSession && run.settled) {
-				this._deletedRlmChildIds.add(childId);
-				this._removeRlmSubagentTracking(childId, run);
-				return { subagent };
-			}
-			if (liveSession && run.settled) {
-				run.deletionRunFinished = true;
-				run.settlement = createAgentMessageDeferred();
-				run.settled = false;
-				this._unsettledRlmChildRuns.add(run);
-			}
-			if (liveSession) this._continueFinishedRlmRunDeletion(run, subagent, liveSession);
-
-			// Return once deletion is accepted. The run stays hidden but unsettled until
-			// abort-insensitive model/tool work unwinds and the shared cleanup finishes.
-			this._deletedRlmChildIds.add(childId);
-			return { subagent };
-		}
-
-		this._emitRlmSubagentRemoval(subagent);
-		const retained = this._rlmChildSessions.get(childId)?.session;
-		try {
-			await this._deleteRlmSubagentSession(childId, retained);
-		} catch (error) {
-			if (this._disposed || this._disposing) {
-				this._removeRlmSubagentTracking(childId);
-				void retained?.disposeAsync().catch(() => undefined);
-			} else {
-				this._rlmChildCleanupFailures.set(childId, subagent);
-			}
-			throw error;
-		}
-		this._deletedRlmChildIds.add(childId);
-		this._removeRlmSubagentTracking(childId);
-		return { subagent };
+	deleteRlmSubagent(target: string): Promise<RlmDeleteSubagentResult> {
+		return this._children.deleteRlmSubagent(target);
 	}
 
 	/**
@@ -6747,174 +5977,25 @@ export class AgentSession {
 	 * matching event forwarder too.
 	 */
 	registerRlmChildSession(childId: string, session: AgentSession, unsubscribe?: () => void): boolean {
-		// A child can finish concurrently while the parent is (or has) torn down; don't
-		// resurrect the map (it would never be disposed), just drop the child now.
-		if (this._deletingRlmChildren.has(childId) || this._deletedRlmChildIds.has(childId)) {
-			return false;
-		}
-		if (this._subagentRuntimeHost?.completeRlmSubagentRuntime?.(childId, session) === false) {
-			return false;
-		}
-		if (this._disposed || this._disposing) {
-			void session.disposeAsync().catch(() => undefined);
-			return false;
-		}
-		this._rlmChildSessions.set(childId, { session, run: this._activeRlmChildRuns.get(childId) });
-		if (unsubscribe) {
-			this._rlmChildUnsubscribes.set(childId, unsubscribe);
-		}
-		return true;
+		return this._children.registerRlmChildSession(childId, session, unsubscribe);
 	}
 
 	releaseRlmChildSession(childId: string, session: AgentSession): (() => void) | false {
-		const run = this._activeRlmChildRuns.get(childId);
-		if (run?.session === session && run.status === "done") {
-			const unsubscribe = run.unsubscribe ?? noopRlmChildEventUnsubscribe;
-			return () => {
-				run.unsubscribe = undefined;
-				this._activeRlmChildRuns.delete(childId);
-				unsubscribe();
-			};
-		}
-		if (this._rlmChildSessions.get(childId)?.session !== session) return false;
-		const unsubscribe = this._rlmChildUnsubscribes.get(childId) ?? noopRlmChildEventUnsubscribe;
-		return () => {
-			this._rlmChildUnsubscribes.delete(childId);
-			this._rlmChildSessions.delete(childId);
-			unsubscribe();
-		};
-	}
-
-	private _rlmChildSnapshotForRun(
-		run: RlmChildRun,
-		child = run.session ?? this._rlmChildSessions.get(run.id)?.session,
-	): RlmChildAgentSnapshot {
-		const model = child?.model ?? run.model;
-		return {
-			id: run.id,
-			parentId: this._rlmParentNodeId,
-			sessionName: child?.sessionName ?? run.sessionName,
-			model: `${model.provider}/${model.id}`,
-			label: rlmChildLabel(run.prompt),
-			status: run.status,
-			durationMs: run.durationMs,
-			answerPreview: run.answerPreview,
-			toolUseCount: run.toolUseCount > 0 ? run.toolUseCount : undefined,
-			tokenCount: child?._contextTokensForCurrentMessages(),
-			recap: child?.getCurrentRecap(),
-			sessionDir: run.sessionDir,
-			activity: run.activity,
-			repliedSinceTask: child?._repliedToParentSinceTask,
-			error: run.error,
-		};
-	}
-
-	private _rlmChildSnapshotForSession(childId: string, child: AgentSession): RlmChildAgentSnapshot {
-		let answerPreview: string | undefined;
-		let toolUseCount = 0;
-		const messages =
-			child.state.streamingMessage?.role === "assistant"
-				? [...child.messages, child.state.streamingMessage]
-				: child.messages;
-		for (const message of messages) {
-			if (message.role !== "assistant") continue;
-			const text = compactRlmText(readAssistantText(message));
-			if (text) answerPreview = text;
-			toolUseCount += message.content.filter((block) => block.type === "toolCall").length;
-		}
-		return {
-			id: childId,
-			parentId: this._rlmParentNodeId,
-			sessionName: child.sessionName,
-			model: child.model ? `${child.model.provider}/${child.model.id}` : undefined,
-			label: child.sessionName ?? "child agent",
-			status: "done",
-			answerPreview,
-			toolUseCount: toolUseCount > 0 ? toolUseCount : undefined,
-			tokenCount: child._contextTokensForCurrentMessages(),
-			recap: child.getCurrentRecap(),
-			sessionDir: child._rlmSessionDir ?? child.sessionManager.getSessionDir(),
-			// No run exists (e.g. a child rehydrated after daemon recovery), so live
-			// session state is the only source for in-flight follow-up work. Mirror
-			// the run projection's convention: status stays "done" (the recorded task
-			// finished) and current work surfaces through activity.
-			activity: child.isSessionActive ? { kind: child.isStreaming ? "writing" : "waiting" } : undefined,
-			repliedSinceTask: child._repliedToParentSinceTask,
-		};
-	}
-
-	private _isUnboundTerminalRlmChildRun(run: RlmChildRun): boolean {
-		if (run.session !== undefined || this._rlmChildSessions.has(run.id)) return false;
-		return run.status === "done" || run.status === "error" || run.status === "cancelled";
+		return this._children.releaseRlmChildSession(childId, session);
 	}
 
 	/** Live recursive child roster from lifecycle state, including nested work under retained parents. */
 	getRlmChildSnapshots(): RlmChildAgentSnapshot[] {
-		const snapshots: RlmChildAgentSnapshot[] = [];
-		const recorded = new Set<string>();
-		const traversed = new Set<string>();
-		for (const run of this._activeRlmChildRuns.values()) {
-			const hidden =
-				run.detachedDeletion ||
-				this._deletingRlmChildren.has(run.id) ||
-				this._deletedRlmChildIds.has(run.id) ||
-				this._isUnboundTerminalRlmChildRun(run);
-			const child = run.session;
-			if (!hidden) {
-				snapshots.push(this._rlmChildSnapshotForRun(run));
-				recorded.add(run.id);
-			}
-			if (child) {
-				traversed.add(run.id);
-				snapshots.push(...child.getRlmChildSnapshots());
-			}
-		}
-		for (const [childId, { session: child, run }] of this._rlmChildSessions) {
-			if (recorded.has(childId) || traversed.has(childId)) continue;
-			const hidden = this._deletingRlmChildren.has(childId) || this._deletedRlmChildIds.has(childId);
-			if (!hidden) {
-				const snapshot = run
-					? this._rlmChildSnapshotForRun(run, child)
-					: this._rlmChildSnapshotForSession(childId, child);
-				snapshots.push({
-					...snapshot,
-					status: this._rlmChildCleanupFailures.has(childId) ? "cancelled" : snapshot.status,
-				});
-			}
-			snapshots.push(...child.getRlmChildSnapshots());
-		}
-		return snapshots;
+		return this._children.getRlmChildSnapshots();
 	}
 
 	/** True when any direct or nested subagent is still running or queued. */
 	hasRunningRlmChildren(): boolean {
-		for (const session of this._rlmSubtreeSessions()) {
-			for (const run of session._activeRlmChildRuns.values()) {
-				if (run.status === "running" || run.status === "queued") {
-					return true;
-				}
-			}
-		}
-		return false;
-	}
-
-	private _rlmChildSessionSnapshot(): AgentSession[] {
-		const sessions = new Set<AgentSession>();
-		for (const [childId, { session }] of this._rlmChildSessions) {
-			if (!this._abandonedRlmQuiescenceChildIds.has(childId)) sessions.add(session);
-		}
-		for (const run of this._activeRlmChildRuns.values()) {
-			if (run.session && !run.abandonedForQuiescence) sessions.add(run.session);
-		}
-		return [...sessions];
+		return this._children.hasRunningRlmChildren();
 	}
 
 	private _hasUnsettledRlmQuiescenceWork(): boolean {
-		if (this._hasDeferredRlmTerminalNotices()) return true;
-		if ([...this._unsettledRlmChildRuns].some((run) => !run.settled)) return true;
-		return this._rlmChildSessionSnapshot().some(
-			(child) => child.isSessionActive || child._hasUnsettledRlmQuiescenceWork(),
-		);
+		return this._children.hasUnsettledWork();
 	}
 
 	/**
@@ -6922,60 +6003,13 @@ export class AgentSession {
 	 * message and for the resulting parent turns to drain. Re-snapshotting after
 	 * each drain includes descendants spawned while earlier results were consumed.
 	 */
-	async waitForRlmQuiescence(externalSignal?: AbortSignal): Promise<void> {
-		const cancellation = new AbortController();
-		const cancelFromParent = () => cancellation.abort();
-		if (externalSignal?.aborted) cancellation.abort();
-		else externalSignal?.addEventListener("abort", cancelFromParent, { once: true });
-		this._rlmQuiescenceWaitAborts.add(cancellation);
-		let rejectCancelled = (_error: Error) => {};
-		const cancelled = new Promise<never>((_resolve, reject) => {
-			rejectCancelled = reject;
-		});
-		const onCancelled = () => rejectCancelled(new Error("RLM quiescence wait cancelled"));
-		cancellation.signal.addEventListener("abort", onCancelled, { once: true });
-		if (cancellation.signal.aborted) onCancelled();
-		const wait = <T>(operation: Promise<T>): Promise<T> => Promise.race([operation, cancelled]);
-		try {
-			while (true) {
-				await wait(this.waitForHeadlessIdle());
-				// Strong RLM quiescence also owns work that interactive waitForIdle ignores.
-				if (this.isSessionActive || this._hasDeferredRlmTerminalNotices()) {
-					await wait(this._waitForSessionActivityChange(cancellation.signal));
-					continue;
-				}
-				const unsettledRuns = [...this._unsettledRlmChildRuns].filter((run) => !run.settled);
-				const childSessions = this._rlmChildSessionSnapshot();
-				if (unsettledRuns.length === 0 && !this._hasUnsettledRlmQuiescenceWork()) return;
-				await wait(
-					Promise.all([
-						...unsettledRuns.map((run) => run.settlement.promise),
-						...childSessions.map((child) => child.waitForRlmQuiescence(cancellation.signal)),
-					]),
-				);
-				// Always loop through the self-active/deferred checks again. Work may
-				// start at the child-settlement boundary.
-			}
-		} finally {
-			// A local descendant error must cancel sibling recursive waits owned by
-			// this barrier before their propagation listeners are removed.
-			cancellation.abort();
-			externalSignal?.removeEventListener("abort", cancelFromParent);
-			cancellation.signal.removeEventListener("abort", onCancelled);
-			this._rlmQuiescenceWaitAborts.delete(cancellation);
-		}
+	waitForRlmQuiescence(externalSignal?: AbortSignal): Promise<void> {
+		return this._children.waitForRlmQuiescence(externalSignal);
 	}
 
 	// Inline (non-daemon) mode only; daemon clients attach to the child session directly.
 	getRlmChildSession(childId: string): AgentSession | undefined {
-		for (const session of this._rlmSubtreeSessions()) {
-			const direct =
-				session._activeRlmChildRuns.get(childId)?.session ?? session._rlmChildSessions.get(childId)?.session;
-			if (direct) {
-				return direct;
-			}
-		}
-		return undefined;
+		return this._children.getRlmChildSession(childId);
 	}
 
 	/**
@@ -6985,102 +6019,12 @@ export class AgentSession {
 	 * was suppressed; false when the id is unknown or the run already settled.
 	 */
 	cancelRlmChildRun(childId: string, reason = "Cancelled by user"): boolean {
-		for (const session of this._rlmSubtreeSessions()) {
-			const run = session._activeRlmChildRuns.get(childId);
-			if (run) {
-				if (run.status !== "running" && run.status !== "queued" && !run.settled) {
-					if (session._inputScheduler.suspended) session._abandonRlmRunForQuiescence(run);
-					else run.suppressTerminalNotice = true;
-					return true;
-				}
-				// The abort cascade never reaches running work retained under a settled descendant.
-				const cancelled = session._cancelRlmChildRun(run, reason);
-				const descendantsCancelled = run.session?.cancelRunningRlmDescendants(reason) ?? false;
-				if (cancelled || descendantsCancelled) {
-					return true;
-				}
-			}
-			// A fruitless match keeps walking: child ids are only mkdir-unique among
-			// siblings, so a colliding live run elsewhere must stay reachable.
-			if (session._rlmChildSessions.get(childId)?.session.cancelRunningRlmDescendants(reason)) {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	// A done child sits in BOTH maps until passivation; the visited set keeps that dual membership from doubling the walk.
-	private *_rlmSubtreeSessions(): Generator<AgentSession> {
-		const visited = new Set<AgentSession>([this]);
-		const stack: AgentSession[] = [this];
-		while (stack.length > 0) {
-			const session = stack.pop()!;
-			yield session;
-			for (const run of session._activeRlmChildRuns.values()) {
-				if (run.session && !visited.has(run.session)) {
-					visited.add(run.session);
-					stack.push(run.session);
-				}
-			}
-			for (const { session: retained } of session._rlmChildSessions.values()) {
-				if (!visited.has(retained)) {
-					visited.add(retained);
-					stack.push(retained);
-				}
-			}
-		}
+		return this._children.cancelRlmChildRun(childId, reason);
 	}
 
 	/** Cancel every running or queued run in this session's subtree. */
 	cancelRunningRlmDescendants(reason = "Cancelled by user"): boolean {
-		let cancelled = false;
-		for (const session of this._rlmSubtreeSessions()) {
-			for (const run of session._activeRlmChildRuns.values()) {
-				if (session._cancelRlmChildRun(run, reason)) cancelled = true;
-			}
-		}
-		return cancelled;
-	}
-
-	private async _assertRlmSubagentSessionNameAvailable(name: string, ignorePendingReservation = false): Promise<void> {
-		const depth = this._rlmDepth + 1;
-		if (!ignorePendingReservation && this._pendingRlmSubagentSessionNames.has(name)) {
-			throw new Error(formatAgentSessionNameUnavailable(name, depth));
-		}
-		const localConflict =
-			[...this._activeRlmChildRuns.values()].some(
-				(run) => run.session?.sessionName === name || (!run.session && run.sessionName === name),
-			) ||
-			[...this._rlmChildSessions.values()].some(({ session }) => session.sessionName === name) ||
-			[...this._rlmChildCleanupFailures.values()].some((entry) => entry.session_name === name);
-		if (localConflict) {
-			throw new Error(formatAgentSessionNameUnavailable(name, depth));
-		}
-		const controller = this._agentMessageController;
-		if (!controller) return;
-		const input = {
-			name,
-			depth,
-			parentSessionId: this.sessionId,
-			parentSessionPath: this.sessionFile,
-		};
-		if (controller.assertSessionNameAvailable) {
-			await controller.assertSessionNameAvailable(input);
-			return;
-		}
-		const listed = await controller.listAgents();
-		const catalog = listed.agents.map(
-			(agent): AgentFamilyCatalogEntry => ({
-				id: agent.sessionId,
-				...(agent.sessionName ? { name: agent.sessionName } : {}),
-				depth: agent.rlmDepth ?? 0,
-				status: agent.status ?? "idle",
-				...(agent.parentSessionId ? { parentSessionId: agent.parentSessionId } : {}),
-				...(agent.parentSessionPath ? { parentSessionPath: agent.parentSessionPath } : {}),
-				...(agent.sessionPath ? { sessionPath: agent.sessionPath } : {}),
-			}),
-		);
-		assertAgentSessionNameAvailable(catalog, input);
+		return this._children.cancelRunningRlmDescendants(reason);
 	}
 
 	private async _authenticatedRlmModels(): Promise<Model<Api>[]> {
@@ -7126,538 +6070,8 @@ export class AgentSession {
 		return { model };
 	}
 
-	private async _startRlmChildRun(
-		prompt: string,
-		kwargs: Record<string, unknown> = {},
-		spawnCode?: string,
-	): Promise<RlmSpawnHandle> {
-		// Snapshot before any await: the spawning request is the turn whose tool call is
-		// executing now. A spawn arriving outside an active run (a detached kernel task
-		// firing while the parent is idle) has no such turn; an absent edge beats a wrong one.
-		const spawnedByRequestId = this.isStreaming ? this._semanticEdges.lastTurnRequestId : undefined;
-		const { name: rawName, model: rawModel, thinking: rawThinking, ...unsupported } = kwargs;
-		const unsupportedKwargs = Object.keys(unsupported);
-		if (unsupportedKwargs.length > 0) {
-			throw new Error(`Unsupported rlm.spawn kwargs: ${unsupportedKwargs.sort().join(", ")}`);
-		}
-		const requestedSessionName = normalizeRequestedRlmSubagentSessionName(rawName);
-		const requestedModel = normalizeRequestedRlmSubagentModel(rawModel);
-		const requestedThinkingLevel = normalizeRequestedRlmSubagentThinkingLevel(rawThinking);
-		if (requestedSessionName) assertDirectAgentMessageTarget(requestedSessionName);
-		if (this._rlmDepth >= this._rlmMaxDepth) {
-			throw new Error(
-				`RLM recursion depth limit reached (RLM_DEPTH=${this._rlmDepth}, RLM_MAX_DEPTH=${this._rlmMaxDepth})`,
-			);
-		}
-		if (requestedSessionName) {
-			if (this._pendingRlmSubagentSessionNames.has(requestedSessionName)) {
-				throw new Error(formatAgentSessionNameUnavailable(requestedSessionName, this._rlmDepth + 1));
-			}
-			this._pendingRlmSubagentSessionNames.add(requestedSessionName);
-		}
-		let modelSelection: RlmSubagentModelSelection;
-		try {
-			if (requestedSessionName) await this._assertRlmSubagentSessionNameAvailable(requestedSessionName, true);
-			modelSelection = await this._resolveRlmSubagentModel(requestedModel);
-		} finally {
-			if (requestedSessionName) this._pendingRlmSubagentSessionNames.delete(requestedSessionName);
-		}
-		if (requestedThinkingLevel !== undefined) {
-			const supported = getSupportedThinkingLevels(modelSelection.model) as ThinkingLevel[];
-			if (!supported.includes(requestedThinkingLevel)) {
-				throw new Error(
-					`Requested thinking level "${requestedThinkingLevel}" is not supported by model "${modelSelection.model.provider}/${modelSelection.model.id}"; supported levels: ${supported.join(", ")}`,
-				);
-			}
-		}
-		if (this._disposed || this._disposing) throw new Error("Cannot spawn a subagent after its parent was disposed");
-
-		const childSessionDir = this._createChildRlmSessionDir();
-		const childNodeId = basename(childSessionDir);
-		const sessionName = requestedSessionName ?? createDefaultRlmSubagentSessionName(prompt, childNodeId);
-		if (!requestedSessionName) await this._assertRlmSubagentSessionNameAvailable(sessionName);
-		const startedAt = Date.now();
-		const parentAssistantForUsage = this._findLastAssistantMessage();
-		if (parentAssistantForUsage && !this._rlmDurableParentUsage.has(parentAssistantForUsage)) {
-			this._rlmDurableParentUsage.set(parentAssistantForUsage, cloneUsage(parentAssistantForUsage.usage));
-		}
-		// Child completions accumulate per origin and flush one durable entry per
-		// settle boundary (agent_end, settlement); the staleness checkpoints and
-		// timer bound crash loss to one window of accumulated usage.
-		const pendingChildUsage = new Map<ChildUsageAttributionEntry["origin"], Usage>();
-		let pendingChildUsageSince = 0;
-		let pendingChildUsageTimer: ReturnType<typeof setTimeout> | undefined;
-		let parentEntryDrainScheduled = false;
-		const flushPendingChildUsageAttribution = (afterParentDrain = false) => {
-			if (pendingChildUsageTimer !== undefined) {
-				clearTimeout(pendingChildUsageTimer);
-				pendingChildUsageTimer = undefined;
-			}
-			if (pendingChildUsage.size === 0 || !parentAssistantForUsage) return;
-			const parentEntry = this._findAssistantEntryForMessage(parentAssistantForUsage);
-			if (!parentEntry) {
-				if (!afterParentDrain && !parentEntryDrainScheduled) {
-					parentEntryDrainScheduled = true;
-					const flushAfterParentDrain = () => {
-						parentEntryDrainScheduled = false;
-						flushPendingChildUsageAttribution(true);
-					};
-					// A message_end extension may still be holding the parent assistant before its append.
-					// The parent drain owns this retry; child settlement never waits for that queue.
-					this._agentEventQueue = this._agentEventQueue.then(flushAfterParentDrain, flushAfterParentDrain);
-					this._agentEventQueue.catch(() => {});
-				}
-				return;
-			}
-			const batches = [...pendingChildUsage.entries()];
-			pendingChildUsage.clear();
-			for (const [origin, childUsage] of batches) {
-				const aggregateUsage = cloneUsage(this._rlmDurableParentUsage.get(parentAssistantForUsage)!);
-				attributeChildUsage(aggregateUsage, childUsage);
-				const liveUsage = parentAssistantForUsage.usage;
-				const entryCount = this.sessionManager.getEntries().length;
-				try {
-					this.sessionManager.appendChildUsageAttribution(parentEntry.id, childUsage, aggregateUsage, origin);
-					this._rlmDurableParentUsage.set(parentAssistantForUsage, aggregateUsage);
-				} catch {
-					// Attribution is recoverable bookkeeping; a failed append must not break run settlement.
-				} finally {
-					// The manager updates this same message; retain siblings' still-pending live usage.
-					parentAssistantForUsage.usage = liveUsage;
-					const indexed = this.sessionManager.getEntries()[entryCount];
-					const unindexedUsage = this._rlmUnindexedChildUsage.get(parentAssistantForUsage);
-					// _persist can throw after indexing. That row already participates in live own-usage subtraction.
-					if (
-						indexed?.type === "child_usage_attributed" &&
-						indexed.targetId === parentEntry.id &&
-						unindexedUsage
-					) {
-						subtractAssistantUsage(unindexedUsage, childUsage);
-					}
-					this._ownUsageMemo = undefined;
-				}
-			}
-		};
-		const flushPendingChildUsageIfStale = () => {
-			if (
-				pendingChildUsage.size > 0 &&
-				Date.now() - pendingChildUsageSince >= RLM_CHILD_USAGE_FLUSH_MAX_PENDING_MS
-			) {
-				flushPendingChildUsageAttribution();
-			}
-		};
-		let runningToolCount = 0;
-		let childSession: AgentSession | undefined;
-		const run: RlmChildRun = {
-			id: childNodeId,
-			prompt,
-			sessionName,
-			sessionDir: childSessionDir,
-			model: modelSelection.model,
-			status: "queued",
-			toolUseCount: 0,
-			settled: false,
-			abort: noopRlmChildAbort,
-			publication: createAgentMessageDeferred(),
-			settlement: createAgentMessageDeferred(),
-			deletionReservation: createAgentMessageDeferred(),
-		};
-		const throwIfCancelled = () => {
-			if (run.status === "cancelled") throw new Error(run.error ?? "RLM child cancelled");
-		};
-		this._activeRlmChildRuns.set(run.id, run);
-		this._unsettledRlmChildRuns.add(run);
-		const emitChildUpdate = () => {
-			const child = this._rlmChildSnapshotForRun(run);
-			const serialized = JSON.stringify(child);
-			if (serialized === run.lastEmittedUpdate) return;
-			run.lastEmittedUpdate = serialized;
-			this._emit({ type: "rlm_child_update", child });
-		};
-		run.emitUpdate = emitChildUpdate;
-		emitChildUpdate();
-
-		const publishChildSession = (child: AgentSession) => {
-			childSession = child;
-			if (this._activeRlmChildRuns.get(run.id) !== run) return;
-			run.session = child;
-			run.abort = () => void child.abort();
-			run.publication.resolve();
-			// Cancellation may have been admitted while runtime construction was
-			// blocked and run.abort was still a no-op.
-			if (run.status === "cancelled") run.abort();
-		};
-		const subagentOptions: CreateRlmSubagentRuntimeOptions = {
-			...this._createRlmSubagentRuntimeOptions({
-				id: childNodeId,
-				prompt,
-				sessionName,
-				spawnCode,
-				sessionDir: childSessionDir,
-				model: modelSelection.model,
-				thinkingLevel: requestedThinkingLevel,
-				spawnedByRequestId,
-			}),
-			onSessionPublished: publishChildSession,
-		};
-
-		const deliverTerminalMessageToParent = async (message: CustomMessage): Promise<void> => {
-			// Synthesized lifecycle notices always use the parent's private durable
-			// path. Explicit child replies continue through agent_message separately.
-			await this._deferRlmTerminalNotice(message);
-		};
-
-		run.completeDeletion = () => {
-			if (!run.deletionNeedsCompletionNotice || run.suppressTerminalNotice || this._disposed || this._disposing) {
-				return Promise.resolve();
-			}
-			if (run.deletionNotice) return run.deletionNotice;
-			const notice = deliverTerminalMessageToParent(
-				createRlmChildTerminalNoticeMessage({
-					kind: "cancelled",
-					childId: run.id,
-					sessionName,
-					reason: run.error ?? "Deleted by parent orchestrator",
-				}),
-			);
-			run.deletionNotice = notice;
-			return notice;
-		};
-
-		run.reportDeletionCleanupFailure = (error) => {
-			if (run.suppressTerminalNotice || this._disposed || this._disposing) return Promise.resolve();
-			if (run.deletionFailureNotice) return run.deletionFailureNotice;
-			const cleanupError = error instanceof Error ? error.message : String(error);
-			const notice = deliverTerminalMessageToParent(
-				createRlmChildFailureMessage({
-					childId: run.id,
-					sessionName,
-					error: `Deletion cleanup failed; retry rlm.delete_subagent("${run.id}") before completion: ${cleanupError}`,
-				}),
-			);
-			run.deletionFailureNotice = notice;
-			return notice;
-		};
-
-		// Runtime startup and the task run are deliberately detached. The public
-		// spawn resolves at admission, while this task owns live tracking, usage,
-		// retention, cancellation, and late-startup cleanup.
-		void (async () => {
-			let childRuntime: RlmSubagentRuntime | undefined;
-			try {
-				childRuntime = await this._createRlmSubagentRuntime(subagentOptions);
-				const child = childRuntime.session;
-				if (run.status === "cancelled") throw new Error(run.error ?? "RLM child cancelled");
-				if (child.sessionName !== sessionName) child.setSessionName(sessionName);
-				publishChildSession(child);
-				throwIfCancelled();
-				run.status = "running";
-				emitChildUpdate();
-				const unsubscribeChildEvents = child.subscribe((event) => {
-					if (event.type === "rlm_child_update") {
-						this._emit(event);
-						return;
-					}
-					if (event.type === "agent_start") {
-						run.activity = { kind: "waiting" };
-						emitChildUpdate();
-					} else if (event.type === "agent_end") {
-						flushPendingChildUsageAttribution();
-						run.activity = undefined;
-						emitChildUpdate();
-					} else if (event.type === "message_end" && event.message.role === "assistant") {
-						const assistant = event.message as AssistantMessage;
-						if (assistant.stopReason !== "error" && assistant.stopReason !== "aborted") {
-							// Flush before the fold: a persisted aggregate may only include
-							// completions whose childUsage is durable with or before it.
-							flushPendingChildUsageIfStale();
-							attributeChildUsage(parentAssistantForUsage?.usage ?? emptyUsage(), assistant.usage);
-							if (parentAssistantForUsage) {
-								const unindexedUsage =
-									this._rlmUnindexedChildUsage.get(parentAssistantForUsage) ?? emptyUsage();
-								addAssistantUsage(unindexedUsage, assistant.usage);
-								this._rlmUnindexedChildUsage.set(parentAssistantForUsage, unindexedUsage);
-								this._ownUsageMemo = undefined;
-								const origin = rlmChildUsageOrigin(child.messages, assistant);
-								if (pendingChildUsage.size === 0) {
-									pendingChildUsageSince = Date.now();
-									// Wall-clock backstop for long tool runs without checkpoints.
-									pendingChildUsageTimer = setTimeout(
-										flushPendingChildUsageAttribution,
-										RLM_CHILD_USAGE_FLUSH_MAX_PENDING_MS,
-									);
-									pendingChildUsageTimer.unref?.();
-								}
-								const bucket = pendingChildUsage.get(origin) ?? emptyUsage();
-								addAssistantUsage(bucket, assistant.usage);
-								pendingChildUsage.set(origin, bucket);
-							}
-						}
-						const text = compactRlmText(readAssistantText(assistant));
-						if (text) run.answerPreview = text;
-						emitChildUpdate();
-					} else if (event.type === "message_start" || event.type === "message_update") {
-						if (event.message.role === "assistant") {
-							const text = compactRlmText(readAssistantText(event.message as AssistantMessage));
-							if (text) run.answerPreview = text;
-							run.activity = { kind: "writing" };
-							emitChildUpdate();
-						}
-					} else if (event.type === "tool_execution_start") {
-						flushPendingChildUsageIfStale();
-						run.toolUseCount += 1;
-						runningToolCount += 1;
-						run.activity = { kind: "executing", toolName: event.toolName };
-						emitChildUpdate();
-					} else if (event.type === "tool_execution_end") {
-						runningToolCount = Math.max(0, runningToolCount - 1);
-						if (runningToolCount === 0) run.activity = { kind: "waiting" };
-						emitChildUpdate();
-					} else if (event.type === "session_info_changed" || event.type === "recap_update") {
-						emitChildUpdate();
-					}
-				});
-				run.unsubscribe = unsubscribeChildEvents;
-				const content = `[task from parent]\n\n${prompt}`;
-				const spawnMessage: AgentSessionMessage = {
-					role: "custom",
-					customType: AGENT_MESSAGE_CUSTOM_TYPE,
-					content,
-					display: true,
-					details: {
-						id: `spawn:${run.id}`,
-						message: prompt,
-						from: {
-							sessionId: this.sessionId,
-							sessionName: this.sessionName,
-							activeSessionId: await this._currentActiveSessionId(),
-						},
-						fromRelationship: "parent",
-					},
-					timestamp: Date.now(),
-				};
-				throwIfCancelled();
-				const parentReplyCountBeforeRun = child._parentReplyCount;
-				await child.promptAndWait(content, {
-					expandPromptTemplates: false,
-					source: "extension",
-					customMessage: spawnMessage,
-				});
-				await child.waitForRlmQuiescence();
-				if (run.error) throw new Error(run.error);
-				run.status = "done";
-				// Only successful completions return; the edge lands on the parent's next commit.
-				const childLastCommitted = child.semanticEdges.lastCommittedRequestId;
-				if (childLastCommitted !== undefined) {
-					this._semanticEdges.recordChildReturned(child.sessionId, childLastCommitted);
-				}
-				run.durationMs = Date.now() - startedAt;
-				run.activity = undefined;
-				emitChildUpdate();
-				if (
-					!run.detachedDeletion &&
-					!run.suppressTerminalNotice &&
-					child._parentReplyCount === parentReplyCountBeforeRun
-				) {
-					const lastAssistantText = child.getLastAssistantText();
-					await deliverTerminalMessageToParent(
-						createRlmChildTerminalNoticeMessage({
-							kind: "completed_without_reply",
-							childId: run.id,
-							sessionName,
-							lastAssistantTextPreview: lastAssistantText ? compactRlmText(lastAssistantText) : undefined,
-						}),
-					);
-				}
-				if (!this.registerRlmChildSession(run.id, child) && !run.detachedDeletion) {
-					if (childRuntime && this._subagentRuntimeHost?.releaseRlmSubagentRuntime) {
-						await this._subagentRuntimeHost
-							.releaseRlmSubagentRuntime(childRuntime, subagentOptions, "error")
-							.catch(() => void child.disposeAsync().catch(() => undefined));
-					} else {
-						await child.disposeAsync().catch(() => undefined);
-					}
-				}
-			} catch (error) {
-				const runError = error instanceof Error ? error : new Error(String(error));
-				run.publication.reject(runError);
-				if (run.status !== "cancelled") {
-					run.status = "error";
-					run.error = runError.message;
-				}
-				// A failed child still returns an error outcome the parent consumes;
-				// cancelled runs and zero-commit children return nothing.
-				const failedChild = childSession ?? childRuntime?.session;
-				const failedLastCommitted = failedChild?.semanticEdges.lastCommittedRequestId;
-				if (run.status === "error" && failedChild && failedLastCommitted !== undefined) {
-					this._semanticEdges.recordChildReturned(failedChild.sessionId, failedLastCommitted);
-				}
-				run.durationMs = Date.now() - startedAt;
-				run.activity = undefined;
-				if (run.status === "error" && childSession === undefined) {
-					// A pre-bind failure leaves no row: "cancelled" is the wire's removal signal.
-					this._emit({
-						type: "rlm_child_update",
-						child: { ...this._rlmChildSnapshotForRun(run), status: "cancelled" },
-					});
-				} else {
-					emitChildUpdate();
-				}
-				if (!run.detachedDeletion && !run.suppressTerminalNotice) {
-					if (run.status === "error") {
-						await deliverTerminalMessageToParent(
-							createRlmChildFailureMessage({
-								childId: run.id,
-								sessionName,
-								error: run.error ?? "unknown error",
-							}),
-						);
-					} else if (run.status === "cancelled") {
-						await deliverTerminalMessageToParent(
-							createRlmChildTerminalNoticeMessage({
-								kind: "cancelled",
-								childId: run.id,
-								sessionName,
-								reason: run.error,
-							}),
-						);
-					}
-				}
-				if (!run.detachedDeletion && childSession && this._subagentRuntimeHost?.releaseRlmSubagentRuntime) {
-					try {
-						await this._subagentRuntimeHost.releaseRlmSubagentRuntime(
-							childRuntime ?? { session: childSession },
-							subagentOptions,
-							run.status === "cancelled" ? "cancelled" : "error",
-						);
-						if (run.status === "cancelled" && !this._disposed && !this._disposing) {
-							this._deletedRlmChildIds.add(run.id);
-							this._removeRlmSubagentTracking(run.id);
-						}
-					} catch {
-						await childSession?.disposeAsync().catch(() => undefined);
-					}
-				} else if (!run.detachedDeletion) {
-					try {
-						if (childRuntime && this._subagentRuntimeHost) {
-							await this._subagentRuntimeHost.deleteRlmSubagentRuntime(run.id, childRuntime.session);
-						} else if (childSession) {
-							await childSession.disposeAsync();
-						}
-						if (run.status === "cancelled" && !this._disposed && !this._disposing) {
-							this._deletedRlmChildIds.add(run.id);
-							this._removeRlmSubagentTracking(run.id);
-						}
-					} catch {
-						// A failed best-effort retry remains available through the retained cleanup maps.
-					}
-				}
-			} finally {
-				flushPendingChildUsageAttribution();
-				if (run.detachedDeletion) {
-					run.deletionRunFinished = true;
-					if (!run.settled) {
-						let cleanupSucceeded = !run.deletionCleanupFailed;
-						if (childRuntime && cleanupSucceeded) {
-							const cleanup =
-								run.deletionCleanup ?? this._ensureRlmRunDeletionCleanup(run, childRuntime.session);
-							cleanupSucceeded = await this._observeRlmRunDeletionCleanup(
-								run,
-								run.detachedDeletion,
-								childRuntime.session,
-								cleanup,
-							);
-						}
-						if (cleanupSucceeded) await this._finishRlmRunDeletion(run);
-					}
-				} else {
-					if (this._activeRlmChildRuns.get(run.id) === run) {
-						if (this._rlmChildSessions.has(run.id)) {
-							this._activeRlmChildRuns.delete(run.id);
-							if (run.unsubscribe) this._rlmChildUnsubscribes.set(run.id, run.unsubscribe);
-							run.abort = noopRlmChildAbort;
-							run.unsubscribe = undefined;
-							run.session = undefined;
-						} else if (run.status !== "error") {
-							this._removeRlmSubagentTracking(run.id, run);
-						} else {
-							run.unsubscribe?.();
-							run.abort = noopRlmChildAbort;
-							run.unsubscribe = undefined;
-						}
-					}
-					run.settled = true;
-					run.settlement.resolve();
-					this._unsettledRlmChildRuns.delete(run);
-					this._maybeResumeGoalContinuationAfterRlmWork();
-				}
-			}
-		})().catch(() => undefined);
-
-		return {
-			rlm_child_id: childNodeId,
-			name: sessionName,
-			session_dir: childSessionDir,
-			model: `${modelSelection.model.provider}/${modelSelection.model.id}`,
-		};
-	}
-
-	async createRlmSession(prompt: string, kwargs: Record<string, unknown> = {}): Promise<RlmCreateSessionResult> {
-		const { name: rawName, model: rawModel, thinking: rawThinking, cwd: rawCwd, ...unsupported } = kwargs;
-		const unsupportedKeys = Object.keys(unsupported);
-		if (unsupportedKeys.length > 0) {
-			throw new Error(`Unsupported rlm.create_session kwargs: ${unsupportedKeys.sort().join(", ")}`);
-		}
-		if (!prompt.trim()) {
-			throw new Error("rlm.create_session prompt must not be empty");
-		}
-		if (this._rlmDepth !== 0) {
-			throw new Error("rlm.create_session is available only from a depth-0 session");
-		}
-		if (this._disposed || this._disposing) {
-			throw new Error("Cannot create a top-level session after the current session was disposed");
-		}
-		const host = this._subagentRuntimeHost;
-		if (!host?.createRlmRootSession) {
-			throw new Error("rlm.create_session requires a daemon-backed depth-0 session");
-		}
-
-		const operation = "rlm.create_session";
-		const sessionName = normalizeRequestedRlmSubagentSessionName(rawName, operation);
-		const requestedModel = normalizeRequestedRlmSubagentModel(rawModel, operation);
-		const requestedThinkingLevel = normalizeRequestedRlmSubagentThinkingLevel(rawThinking, operation);
-		if (sessionName) {
-			assertDirectAgentMessageTarget(sessionName);
-			const controller = this._agentMessageController;
-			if (controller?.assertSessionNameAvailable) {
-				await controller.assertSessionNameAvailable({ name: sessionName, depth: 0 });
-			}
-		}
-		if (rawCwd !== undefined && (typeof rawCwd !== "string" || !rawCwd.trim())) {
-			throw new Error("rlm.create_session cwd must be a non-empty string");
-		}
-		const cwd = rawCwd === undefined ? this._cwd : resolve(this._cwd, rawCwd.trim());
-		const modelSelection = await this._resolveRlmSubagentModel(requestedModel, "top-level session");
-		if (requestedThinkingLevel !== undefined) {
-			const supported = getSupportedThinkingLevels(modelSelection.model) as ThinkingLevel[];
-			if (!supported.includes(requestedThinkingLevel)) {
-				throw new Error(
-					`Requested thinking level "${requestedThinkingLevel}" is not supported by model "${modelSelection.model.provider}/${modelSelection.model.id}"; supported levels: ${supported.join(", ")}`,
-				);
-			}
-		}
-		const thinkingLevel =
-			requestedThinkingLevel ?? (clampThinkingLevel(modelSelection.model, this.thinkingLevel) as ThinkingLevel);
-		if (this._disposed || this._disposing) {
-			throw new Error("Cannot create a top-level session after the current session was disposed");
-		}
-		return host.createRlmRootSession({
-			prompt,
-			sessionName,
-			cwd,
-			model: modelSelection.model,
-			thinkingLevel,
-		});
+	createRlmSession(prompt: string, kwargs: Record<string, unknown> = {}): Promise<RlmCreateSessionResult> {
+		return this._children.createRlmSession(prompt, kwargs);
 	}
 
 	async runRlmChild(
@@ -7665,7 +6079,7 @@ export class AgentSession {
 		kwargs: Record<string, unknown> = {},
 		spawnCode?: string,
 	): Promise<RlmSpawnHandle> {
-		return this._startRlmChildRun(prompt, kwargs, spawnCode);
+		return this._children.run(prompt, kwargs, spawnCode);
 	}
 
 	abortRetry(): void {
@@ -7736,39 +6150,11 @@ export class AgentSession {
 	}
 
 	getRlmMaxDepthStatus(): RlmMaxDepthStatus {
-		return { maxDepth: this._rlmMaxDepth, source: this._rlmMaxDepthSource };
+		return this._childState.getRlmMaxDepthStatus();
 	}
 
-	async setRlmMaxDepth(maxDepth: number, options: { global?: boolean } = {}): Promise<SetRlmMaxDepthResult> {
-		if (!isNonNegativeInteger(maxDepth)) {
-			throw new Error("RLM max depth must be a non-negative integer.");
-		}
-
-		this.sessionManager.appendCustomEntryWithRollback(RLM_MAX_DEPTH_STATE_CUSTOM_TYPE, { maxDepth });
-		this._rlmMaxDepth = maxDepth;
-		this._rlmMaxDepthSource = "chat";
-		const oldBase = this._baseSystemPrompt;
-		this._baseSystemPrompt = this._rebuildSystemPrompt(this.getActiveToolNames());
-		this.agent.state.systemPrompt = this._refreshExtensionSystemPrompt(this.agent.state.systemPrompt, oldBase);
-
-		let globalError: string | undefined;
-		if (options.global) {
-			await this.settingsManager.flush();
-			const staleErrors = this.settingsManager.drainErrors("global");
-			for (const { error } of staleErrors) {
-				console.warn(`Warning: Earlier global settings write failed: ${error.message}`);
-			}
-			this.settingsManager.setRlmMaxDepth(maxDepth);
-			await this.settingsManager.flush();
-			const errors = this.settingsManager.drainErrors("global");
-			globalError = errors.map(({ error }) => error.message).join("; ") || undefined;
-		}
-
-		return {
-			...this.getRlmMaxDepthStatus(),
-			globalSaved: options.global === true && globalError === undefined,
-			...(globalError ? { globalError } : {}),
-		};
+	setRlmMaxDepth(maxDepth: number, options: { global?: boolean } = {}): Promise<SetRlmMaxDepthResult> {
+		return this._childState.setRlmMaxDepth(maxDepth, options);
 	}
 
 	setSessionName(name: string): void {
@@ -8179,11 +6565,17 @@ export class AgentSession {
 	private _subtractUnindexedChildUsage(ownUsage: Usage, entries: SessionEntry[]): void {
 		for (const entry of entries) {
 			if (entry.type !== "message" || entry.message.role !== "assistant") continue;
-			const unindexedUsage = this._rlmUnindexedChildUsage.get(entry.message);
-			if (unindexedUsage) subtractAssistantUsage(ownUsage, unindexedUsage);
+			const unindexed = this._getUnindexedChildUsage(entry.message);
+			if (unindexed) subtractAssistantUsage(ownUsage, unindexed);
 		}
 	}
 
+	private _invalidateOwnUsage(): void {
+		this._ownUsageMemo = undefined;
+	}
+	private _getUnindexedChildUsage(message: AssistantMessage): Usage | undefined {
+		return this._childUsage.getUnindexed(message);
+	}
 	private _ownUsageMemo?: { count: number; tailId: string | undefined; usage: SessionUsageSummary | undefined };
 
 	// Whole-file own spend, identical to the catalog scan so rows never shift at passivation.
@@ -8215,7 +6607,7 @@ export class AgentSession {
 
 		const children: ContextTreeNode[] = [];
 		const liveIds = new Set<string>();
-		for (const run of this._activeRlmChildRuns.values()) {
+		for (const run of this._children.getActiveRuns()) {
 			liveIds.add(run.id);
 			const node =
 				run.session?.getContextTree() ?? loadContextTreeChildFromDisk(run.sessionDir, resolveContextWindow);
